// src/entities/InvoiceDetail/domain/interfaces/IInvoiceDetailController.ts
import type { Request, Response } from "express";

export interface IInvoiceDetailController {
  get: (req: Request, res: Response) => Promise<void>;
  getAll: (req: Request, res: Response) => Promise<void>;
  post: (req: Request, res: Response) => Promise<void>;
  patch: (req: Request, res: Response) => Promise<void>;
  delete: (req: Request, res: Response) => Promise<void>;
}